Add cmake build
