+++ .github/workflows/ccpp.yml
@@ -1,0 +1,58 @@
+name: C/C++ CI
+
+on: [push]
+
+jobs:
+  build-ubuntu:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        config: [Debug, Release]
+        standard: [11, 17]
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Build & Test
+      run: |
+        cmake -E remove_directory build
+        cmake -B build -S . -DCMAKE_BUILD_TYPE=${{ matrix.config }} -DCMAKE_CXX_STANDARD=${{ matrix.standard }} -DCMAKE_CXX_FLAGS="-Werror -fsanitize=address,undefined"
+        cmake --build build
+        cd build
+        ctest --output-on-failure
+        
+  build-windows:
+
+    runs-on: windows-latest
+    strategy:
+      matrix:
+        config: [Debug, Release]
+        standard: [11, 17]
+    
+    steps:
+    - uses: actions/checkout@v1
+    - name: Build & Test
+      run: | 
+        cmake -E remove_directory build
+        cmake -B build -S . -DCMAKE_CXX_STANDARD=${{ matrix.standard }} # -DCMAKE_CXX_FLAGS="/WX"
+        cmake --build build --config ${{ matrix.config }}
+        cd build
+        ctest --output-on-failure
+
+  build-macos:
+
+    runs-on: macOS-latest
+    strategy:
+      matrix:
+        config: [Debug, Release]
+        standard: [11, 17]
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Build & Test
+      run: |
+        cmake -E remove_directory build
+        cmake -B build -S . -DCMAKE_BUILD_TYPE=${{ matrix.config }} -DCMAKE_CXX_STANDARD=${{ matrix.standard }} -DCMAKE_CXX_FLAGS="-Werror -fsanitize=address,undefined"
+        cmake --build build
+        cd build
+        ctest --output-on-failure
